STEP UP and Application Upload
STEP UP FUCK U

STEP UP FUCK U

STEP UP FUCK U

save

save

save

save

Update stim_lib_volt.c

save

fuck

save

FUCK

test

save

save

Update stim_lib_st_init.c

Update stim_lib_st_inc.c

Update stim_lib_st_init.c

Step Up Fix

Step Up Fix

Update stim_lib_st_init.c

Update stim_lib_st_init.c

Update stim_lib_st_init.c

Update main.c

Update stim_lib_st_init.c

save

save

save

Update stim_lib_volt.c

save

Update stim_lib_volt_table.c

Update stim_lib_st_init.c

ADC FUCK U

save

Update main.c

save

save

Update td_stim_mask_param_setting.c

save

save

save

FUCK

save

save

save

save

save

save

save

Update stim_lib_volt.c

Update stim_lib_volt.c

RANK Change

FUCK

save

FUCK

FUCK

fuck

Update td_stim_mask_param_table.c

Update td_stim_mask_param_table.c

GPIO MODE ADD

Update td_schedule.c

Update .cproject

save

Update main.c

Update stim_lib_stim_cfg_st.c

FUCK

save

FUCK

Update stim_lib.c

Update stim_lib.c

FUCK

FUCK

Update stim_lib_stim_cfg_st.c

Update stim_lib_stim_cfg_st.c

save

save

Update main.c

Update td_sys_fsm_state.c

Update main.c

Update td_stim_mask_param_table.c

save

Update td_stim_mask_param_table.c

save

Update main.c

save

Update main.c

+++ 48 PIN/allears_EVK_BULK_48PIN/Core/Src/stim_lib/stim_lib_volt_st.c
@@ -13,8 +13,8 @@
 extern uint16_t adc1_conv_buff[STIM_LIB_ADC1_TOTAL_SIZE]; // ADC1
 
 /* STEPUP CONTROL VALUE :: PULSE and Counter */
-extern uint8_t voltage_ctrl_pulse;
-extern uint8_t stepup_feedback_cnt;
+extern int voltage_ctrl_pulse;
+extern int stepup_feedback_cnt;
 
 void stimLib_stepup_startRaw(void)
 {
